// models/RegisterModel.tsx
export default interface RegisterModel 
{
    nome: string;
    dataNascimento: Date;
    cpf: Number;
    rg: string;
    email: string;
    sexo: Number;
    nomeMae: string;
    senha: string;
    logradouro: string;
    numeroEndereco: Number;
    pais: string;
    uf: string;
    cidade: string;
    bairro: string;
    cep: Number;
    complemento: string;
}